# docs/swagger.yaml
definitions:
  dto.CreateVoucherRequest:
    properties:
      discount_percent:
        maximum: 100
        minimum: 0
        type: number
      expiry_date:
        type: string
      voucher_code:
        description: voucher_code, discount_percent, expiry_date
        maxLength: 255
        type: string
    required:
    - discount_percent
    - expiry_date
    - voucher_code
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.VoucherResponse:
    properties:
      created_at:
        type: string
      discount_percent:
        type: integer
      expiry_date:
        type: string
      id:
        type: string
      updated_at:
        type: string
      voucher_code:
        type: string
    type: object
  util.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing vouchers
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Technical Test Indico API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return token
      parameters:
      - description: Login request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: User login
      tags:
      - Auth
  /vouchers:
    get:
      consumes:
      - application/json
      description: Retrieve a list of vouchers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - default: expiry_date
        description: Sort by field
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order (asc or desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.VoucherResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: List vouchers
      tags:
      - vouchers
    post:
      consumes:
      - application/json
      description: Create a new voucher with the provided details
      parameters:
      - description: Voucher data
        in: body
        name: voucher
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVoucherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.VoucherResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Create a new voucher
      tags:
      - vouchers
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"